requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
imbalanced-learn
astroquery
google-generativeai
joblib
streamlit